tests/requirements.txt
# Agentic Workspace — Test Dependencies
#
# Install: pip install -r tests/requirements.txt

# Core
pytest>=8.0
pytest-asyncio>=0.23

# HTTP testing
httpx>=0.27

# Browser testing (optional — install only if using browser tests)
# pytest-playwright>=0.5
# playwright>=1.40

# Utilities
python-dotenv>=1.0
pyyaml>=6.0
